[Silver V] Title: 타일 장식물, Time: 0 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13301. 타일 장식물/타일 장식물.cc
@@ -8,7 +8,7 @@ int main() {
 	int n;
 	cin >> n;
 
-	int dp[81]{};
+	long long dp[81]{};
 	dp[1] = 4;
 	dp[2] = 6;
